Mech Affinities summary for affinity overflow.
New Features:

Added showRemainingAffinitiesSummary.  Instead of dropping
affinities that are not included in the topAffinitiesInTooltipCount list,
a single line summary showing the chassis name and deployment count.
For example "Griffin (2)"

Added showRemainingAffinitiesSummary in settings for the feature with
the default of disabled.

Changes:

Changed mech affinity list to excluded any mechs that have zero
affinity.

Minor re factor of getPilotToolTip.  Required as the previous code
depended on the chassisValues (which is mech/drop count) to match the
affinities list.

Added showRemainingAffinitiesSummary to docs.